pyproject.toml
[project]
name = "peckin-pairs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
]

[tool.uv]
# This prevents 'uv run' and 'uv sync' from including dev dependencies by default
default-groups = []

[dependency-groups]
dev = [
    "behave>=1.3.3",
    "busypie>=0.6.1",
    "docker>=7.1.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "selenium>=4.41.0",
]
